Renamed configparser.py to config.py
